BOOKING-239: Moved AAKResourceRepository from Main namespace to Resources namespace to match AAKResource entity
