+++ src/Security/Voter/BookingVoter.php
@@ -3,7 +3,7 @@
 namespace App\Security\Voter;
 
 use App\Entity\Main\Booking;
-use App\Repository\Main\AAKResourceRepository;
+use App\Repository\Resources\AAKResourceRepository;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Authorization\Voter\Voter;
